+++ eras/conway/impl/src/Cardano/Ledger/Conway/Rules/Enact.hs
@@ -3,7 +3,6 @@
 {-# LANGUAGE FlexibleInstances #-}
 {-# LANGUAGE MultiParamTypeClasses #-}
 {-# LANGUAGE ScopedTypeVariables #-}
-{-# LANGUAGE TypeApplications #-}
 {-# LANGUAGE TypeFamilies #-}
 {-# LANGUAGE UndecidableInstances #-}
 {-# OPTIONS_GHC -Wno-orphans #-}
@@ -12,12 +11,10 @@ module Cardano.Ledger.Conway.Rules.Enact (
   ConwayENACT,
   EnactSignal (..),
   EnactState (..),
-  EnactPredFailure (..),
 ) where
 
 import Cardano.Ledger.Address (RewardAcnt (..))
 import Cardano.Ledger.BaseTypes
-import Cardano.Ledger.Coin (Coin (..))
 import Cardano.Ledger.Conway.Core
 import Cardano.Ledger.Conway.Era (ConwayENACT)
 import Cardano.Ledger.Conway.Governance (
@@ -34,7 +31,6 @@ import Cardano.Ledger.Conway.Governance (
   ensPrevPParamUpdateL,
   ensProtVerL,
  )
-import Cardano.Ledger.Rules.ValidationMode (Inject (..), runTest)
 import Cardano.Ledger.Val (Val (..))
 import Control.State.Transition.Extended (
   STS (..),
@@ -43,14 +39,9 @@ import Control.State.Transition.Extended (
   judgmentContext,
  )
 import Data.Foldable (fold)
-import Data.Map.Strict (Map)
 import qualified Data.Map.Strict as Map
+import Data.Void (Void)
 import Lens.Micro
-import Validation (failureUnless)
-
-data EnactPredFailure era
-  = EnactTreasuryInsufficientFunds !(Map (RewardAcnt (EraCrypto era)) Coin) !Coin
-  deriving (Show, Eq)
 
 data EnactSignal era = EnactSignal
   { esGovActionId :: !(GovActionId (EraCrypto era))
@@ -59,7 +50,7 @@ data EnactSignal era = EnactSignal
 
 instance EraGov era => STS (ConwayENACT era) where
   type Environment (ConwayENACT era) = ()
-  type PredicateFailure (ConwayENACT era) = EnactPredFailure era
+  type PredicateFailure (ConwayENACT era) = Void
   type Signal (ConwayENACT era) = EnactSignal era
   type State (ConwayENACT era) = EnactState era
   type BaseM (ConwayENACT era) = ShelleyBase
@@ -85,9 +76,6 @@ enactmentTransition = do
     TreasuryWithdrawals wdrls -> do
       let wdrlsAmount = fold wdrls
           wdrlsNoNetworkId = Map.mapKeys getRwdCred wdrls
-      runTest
-        . failureUnless (wdrlsAmount <= ensTreasury st)
-        $ EnactTreasuryInsufficientFunds @era wdrls (ensTreasury st)
       pure
         st
           { ensWithdrawals = Map.unionWith (<>) wdrlsNoNetworkId $ ensWithdrawals st
@@ -109,6 +97,3 @@ enactmentTransition = do
           & ensConstitutionL .~ c
           & ensPrevConstitutionL .~ SJust (PrevGovActionId govActionId)
     InfoAction -> pure st
-
-instance Inject (EnactPredFailure era) (EnactPredFailure era) where
-  inject = id